Rename `destinationPointCheckerCount` to `destinationPointStackIndex`
We are inserted the checker at that index on the stack. The previous name was never correct, as even though we were consistently inserting at the end of the stack, the number we were passing in is equivalent to the last index of the existing checker (i.e., the count of the checkers on a point minus 1).