Allow moving checker for either die for your first move
This means that you can drag to any of the points any die allows you to. Also, clicking on a point will now move to the first die that the checker is allowed to move to. For example, rolling a 1 and 3 from the starting board means that clicking on point 12 causes the element to move to point 15, as point 13 can't be moved to.