Allow undoing of hit checkers and undo checkers with a double roll in the correct order
Before, we were only looking at the value of the dice, with no respect to when it was played temporally.

To fix that issue, and the issue of undoing hit checkers, we now just store the last played moves in a variable, so we can just pop them off the stack. This means that we no longer need to do any complicated logic looking at the current state of the board in order to figure out which checker to undo and by which die value.