Store board state within JS
This somewhat goes against what I've been trying to do up until now, where all the data were stored in data attributes. I unfortunately don't have a full mental model of why this'd be good, so continuing down this route will both be difficult and potentially pointless (if it turns out that this wasn't a good thing to do in the first place).

The motivating factor for moving the board game state into JS was to allow us to have an easy way of accessing the current board state. In the future, we can imagine that this JS representation of backgammon can be pulled out as a library in and of itself, and other people can provide their own backgammon frontends without having to worry about implementing the rules of backgammon.