Fix random HMR failures
Hot module reloading would freeze as soon as it encountered any
compilation error. So when introducing a compilation error (e.g., a
missing semicolon for a CSS rule) and then saving, the HMR is just
frozen. Even if you fix the compilation error by adding the semicolon,
it remains frozen, meaning that I often think that my CSS is wrong, even
when it's not.

What this line of code does is just does a refresh of the page whenever
a change is encountered. Nothing fancy, and it means that it works
predictably.

Similar issue but related to Elm instead of CSS:
https://github.com/parcel-bundler/parcel/issues/6124