Make delay of showing duplicate doubling dice a CSS detail
This means that people that use a11y technology will be aware of the
duplicate doubling dice all at once without having to wait for 200ms.
This also nicely separates the JS and CSS responsibilities, since the
delay in displaying the duplicate doubling dice is purely a visual
queue, and is nothing to do with the core logic of the backgammon game.