Make the doubling cube not a `die`
Few similarities are shared between the die with pips and the doubling
cube. Having them be the same complicates things when we want to convert
the die to a web component, since the doubling cube would need to also
be a `<veyndan-die>`. This means that the pips need to be moved out of
`<veyndan-die>` and replaced with a slot where pips or a number could
go. This is all much more complicated than it needs to be, all to
ultimately save a couple lines of duplication in CSS.