Remove usage of SVG
This was a couple of years in the making (with plenty of breaks).

This breaks dragging of checkers, but instead of fixing the existing code, I'm going to rewrite it as moving off of SVG allows me to use things like flexbox to deal with checkers, removing a bunch of custom maths in the process.